refactor: code split and cleanup
