+++ src/utils/secretReferencing.ts
@@ -1,4 +1,3 @@
-
 import { Secret } from "../types";
 
 type SecretReference = {
@@ -13,7 +12,7 @@ export type SecretFetcher = (
   key: string
 ) => Promise<Secret>;
 
-
+// Regex pattern for secret references
 const REFERENCE_REGEX =
   /\${(?:(?<env>[^.\/}]+)\.)?(?:(?<path>[^}]+)\/)?(?<key>[^}]+)}/g;
 
@@ -21,19 +20,14 @@ export const normalizeKey = (env: string, path: string, key: string) =>
   `${env.toLowerCase()}:${path.replace(/\/+$/, "")}:${key}`;
 
 export function parseSecretReference(reference: string): SecretReference {
-  const parserRegex = new RegExp(REFERENCE_REGEX.source);
-  const match = parserRegex.exec(reference);
-
+  const match = new RegExp(REFERENCE_REGEX.source).exec(reference);
   if (!match?.groups) {
     throw new Error(`Invalid secret reference format: ${reference}`);
   }
 
   let { env, path, key } = match.groups;
-  console.log("reference:", reference, env, path, key);
   env = env?.trim() || "";
   key = key.trim();
-
-  
   path = path ? `/${path.replace(/\.+/g, "/")}`.replace(/\/+/g, "/") : "/";
 
   return { env, path, key };
@@ -48,34 +42,21 @@ export async function resolveSecretReferences(
   resolutionStack: Set<string> = new Set()
 ): Promise<string> {
   const references = Array.from(value.matchAll(REFERENCE_REGEX));
-  console.log("all references", references);
   let resolvedValue = value;
 
   for (const ref of references) {
     try {
-      const {
-        env: refEnv,
-        path: refPath,
-        key: refKey,
-      } = parseSecretReference(ref[0]);
-
+      const { env: refEnv, path: refPath, key: refKey } = parseSecretReference(ref[0]);
       const targetEnv = refEnv || currentEnv;
       const targetPath = refPath || currentPath;
-
       const cacheKey = normalizeKey(targetEnv, targetPath, refKey);
 
-      console.log("util cacheKey", cacheKey);
-
-      console.log("resolving", ref, targetEnv, targetPath, refKey);
-
-      
       if (resolutionStack.has(cacheKey)) {
         throw new Error(`Circular reference detected: ${cacheKey}`);
       }
 
-      // If we already have a resolved value, use it
       if (!cache.has(cacheKey)) {
-        resolutionStack.add(cacheKey); // Add before calling fetcher
+        resolutionStack.add(cacheKey);
         try {
           const secret = await fetcher(targetEnv, targetPath, refKey);
           const resolvedSecretValue = await resolveSecretReferences(
@@ -86,10 +67,9 @@ export async function resolveSecretReferences(
             cache,
             resolutionStack
           );
-
           cache.set(cacheKey, resolvedSecretValue);
         } finally {
-          resolutionStack.delete(cacheKey); // Ensure cleanup even if an error occurs
+          resolutionStack.delete(cacheKey);
         }
       }
